Add annotation type to Darwin v7 metadata to retrieve color
