+++ tests/test_annotations.py
@@ -139,24 +139,27 @@ class TestAnnotations:
     def test_read_darwin_v7(self):
         if not DARWIN_SDK_AVAILABLE:
             return None
-        print(self.v7_annotations.available_classes)
         assert len(self.v7_annotations.available_classes) == 5
-        assert AnnotationClass(label="ROI (segmentation)", annotation_type=AnnotationType.BOX) in self.v7_annotations
-        assert AnnotationClass(label="stroma (area)", annotation_type=AnnotationType.POLYGON) in self.v7_annotations
-        assert AnnotationClass(label="lymphocyte (cell)", annotation_type=AnnotationType.POINT) in self.v7_annotations
-        assert AnnotationClass(label="tumor (cell)", annotation_type=AnnotationType.BOX) in self.v7_annotations
-        assert AnnotationClass(label="tumor (area)", annotation_type=AnnotationType.POLYGON) in self.v7_annotations
-
-        # assert self.v7_annotations.available_classes[0].label == "ROI (segmentation)"
-        # assert self.v7_annotations.available_classes[0].annotation_type == AnnotationType.BOX
-        # assert self.v7_annotations.available_classes[1].label == "stroma (area)"
-        # assert self.v7_annotations.available_classes[1].annotation_type == AnnotationType.POLYGON
-        # assert self.v7_annotations.available_classes[2].label == "lymphocyte (cell)"
-        # assert self.v7_annotations.available_classes[2].annotation_type == AnnotationType.POINT
-        # assert self.v7_annotations.available_classes[3].label == "tumor (cell)"
-        # assert self.v7_annotations.available_classes[3].annotation_type == AnnotationType.BOX
-        # assert self.v7_annotations.available_classes[4].label == "tumor (area)"
-        # assert self.v7_annotations.available_classes[4].annotation_type == AnnotationType.POLYGON
+        assert (
+            AnnotationClass(label="ROI (segmentation)", annotation_type=AnnotationType.BOX, color=(143, 0, 255))
+            in self.v7_annotations
+        )
+        assert (
+            AnnotationClass(label="stroma (area)", annotation_type=AnnotationType.POLYGON, color=(0, 236, 123))
+            in self.v7_annotations
+        )
+        assert (
+            AnnotationClass(label="lymphocyte (cell)", annotation_type=AnnotationType.POINT, color=(0, 236, 123))
+            in self.v7_annotations
+        )
+        assert (
+            AnnotationClass(label="tumor (cell)", annotation_type=AnnotationType.BOX, color=(255, 46, 0))
+            in self.v7_annotations
+        )
+        assert (
+            AnnotationClass(label="tumor (area)", annotation_type=AnnotationType.POLYGON, color=(255, 46, 0))
+            in self.v7_annotations
+        )
 
         assert self.v7_annotations.bounding_box == (
             (15291.49, 18094.48),
@@ -164,7 +167,6 @@ class TestAnnotations:
         )
 
         region = self.v7_annotations.read_region((15300, 19000), 1.0, (2500.0, 2500.0))
-        # FIXME: Is this right? These are the original values
         expected_output = [
             (6250000.0, "BOX", "ROI (segmentation)"),
             (1616768.0657540846, "POLYGON", "stroma (area)"),
@@ -233,7 +235,7 @@ class TestAnnotations:
         annotations = self.asap_annotations.copy()
         annotations.filter(["healthy glands"])
         assert len(annotations._layers) == 1
-        assert annotations.available_classes[0].label == "healthy glands"
+        assert "healthy glands" in annotations
 
         annotations.filter(["non-existing"])
         assert len(annotations._layers) == 0
